finished backend

+++ backend/controller/products.controller.js
@@ -33,13 +33,17 @@ export const getProducts = async(req, res) => {
 export const updateProduct = async(req, res) =>{
     const {id} = req.params;
     const product = req.body;
+
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({success: false, message: "Invalid Product Id"});
+    }
     
     try{
         const updatedProduct =await Product.findByIdAndUpdate(id, product, {new: true});
         res.status(200).json({success: true, data: updatedProduct });
     } catch(error){
         console.error("Error in updating product:", error.message);
-        res.status(500).json({success: false, message: "Server error"});
+        res.status(500).json({success: false, message: "Server Error"});
     } 
 }
 
@@ -47,11 +51,15 @@ export const deleteProduct = async(req, res) => {
     const {ObjectId} = req.params;
     console.log("ObjectId :", ObjectId);
 
+    if(!mongoose.Types.ObjectId.isValid(ObjectId)){
+        return res.status(404).json({success: false, message: "Invalid Product Id"});
+    }
+
     try{
         await Product.findByIdAndDelete(ObjectId);
         res.status(200).json({success: true, message: "Product deleted" });
     } catch(error){
         console.error("Error in deleting product:", error.message);
-        res.status(400).json({success: false, message: "Product not found"});
+        res.status(500).json({success: false, message: "Server Error"});
     }
 }
